+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/DefaultUpdaterResult.cs
@@ -1,0 +1,7 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.DotNet.UpgradeAssistant
+{
+    public record DefaultUpdaterResult(bool Result) : IUpdaterResult { }
+}