+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/FileUpdaterResult.cs
@@ -1,0 +1,9 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+
+namespace Microsoft.DotNet.UpgradeAssistant
+{
+    public record FileUpdaterResult(bool Result, IEnumerable<string> FilePaths) : IUpdaterResult { }
+}