+++ src/steps/Microsoft.DotNet.UpgradeAssistant.Steps.Solution/FinalizeSolutionStep.cs
@@ -29,23 +29,24 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
         protected override Task<UpgradeStepApplyResult> ApplyImplAsync(IUpgradeContext context, CancellationToken token)
         {
             context.IsComplete = true;
-            context.SetEntryPoint(null);
+            context.EntryPoints = Enumerable.Empty<IProject>();
+
             return Task.FromResult(new UpgradeStepApplyResult(UpgradeStepStatus.Complete, "Upgrade complete"));
         }
 
         protected override Task<UpgradeStepInitializeResult> InitializeImplAsync(IUpgradeContext context, CancellationToken token)
         {
-            if (context.EntryPoint is null)
+            if (!context.EntryPoints.Any())
             {
                 return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Complete, "Upgrade completed", BuildBreakRisk.None));
             }
             else
             {
-                return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Incomplete, $"Finalize upgrade of entry point {context.EntryPoint.FileInfo}", BuildBreakRisk.None));
+                return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Incomplete, "Finalize upgrade", BuildBreakRisk.None));
             }
         }
 
         protected override Task<bool> IsApplicableImplAsync(IUpgradeContext context, CancellationToken token)
-            => Task.FromResult(context.CurrentProject is null && context.EntryPoint is not null);
+            => Task.FromResult(context.CurrentProject is null && context.EntryPoints.Any());
     }
 }